requirements.txt
torch==2.5.1
numpy
matplotlib
scipy
PyYAML
statsmodels